Refactor gitignore handling
