+++ options-manager.js
@@ -1,16 +1,14 @@
 'use strict';
-const fs = require('fs');
 const os = require('os');
 const path = require('path');
 const arrify = require('arrify');
 const deepAssign = require('deep-assign');
-const globby = require('globby');
-const gitIgnore = require('ignore');
 const multimatch = require('multimatch');
 const pathExists = require('path-exists');
 const pkgConf = require('pkg-conf');
 const resolveFrom = require('resolve-from');
-const slash = require('slash');
+
+const getGitIgnoreFilter = require('./gitignore').getGitIgnoreFilter;
 
 const DEFAULT_IGNORE = [
 	'**/node_modules/**',
@@ -231,39 +229,6 @@ const groupConfigs = (paths, baseOptions, overrides) => {
 const getIgnores = opts => {
 	opts.ignores = DEFAULT_IGNORE.concat(opts.ignores || []);
 	return opts;
-};
-
-const mapGitIgnorePatternTo = base => ignore => {
-	if (ignore.startsWith('!')) {
-		return '!' + path.posix.join(base, ignore.substr(1));
-	}
-
-	return path.posix.join(base, ignore);
-};
-
-const parseGitIgnore = (content, opts) => {
-	const base = slash(path.relative(opts.cwd, path.dirname(opts.fileName)));
-
-	return content
-		.split(/\r?\n/)
-		.filter(Boolean)
-		.filter(l => l.charAt(0) !== '#')
-		.map(mapGitIgnorePatternTo(base));
-};
-
-const getGitIgnoreFilter = opts => {
-	const ignore = opts.ignores || [];
-	const cwd = opts.cwd || process.cwd();
-
-	const i = globby.sync('**/.gitignore', {ignore, cwd})
-		.reduce((ignores, file) => {
-			const fileName = path.join(cwd, file);
-			const content = fs.readFileSync(fileName, 'utf8');
-			ignores.add(parseGitIgnore(content, {cwd, fileName}));
-			return ignores;
-		}, gitIgnore());
-
-	return p => !i.ignores(slash(path.relative(cwd, p)));
 };
 
 const preprocess = opts => {
