+++ test/options-manager.js
@@ -172,40 +172,6 @@ test('groupConfigs', t => {
 	}));
 });
 
-test('gitignore', t => {
-	const cwd = path.join(__dirname, 'fixtures/gitignore/test');
-	const filter = manager.getGitIgnoreFilter({cwd});
-	const actual = ['foo.js', 'bar.js'].filter(filter);
-	const expected = ['bar.js'];
-	t.deepEqual(actual, expected);
-});
-
-test('ignore ignored .gitignore', t => {
-	const opts = {
-		cwd: path.join(__dirname, 'fixtures/gitignore'),
-		ignores: ['**/test/**']
-	};
-
-	const filter = manager.getGitIgnoreFilter(opts);
-	const actual = ['foo.js'].filter(filter);
-	const expected = ['foo.js'];
-	t.deepEqual(actual, expected);
-});
-
-test.serial('patterns should be translated according to process.cwd()', t => {
-	const previous = process.cwd();
-	const cwd = path.join(__dirname, 'fixtures/gitignore');
-	process.chdir(cwd);
-	try {
-		const filter = manager.getGitIgnoreFilter({});
-		const actual = ['bar.js', 'test/foo.js', 'test/bar.js'].filter(filter);
-		const expected = ['bar.js', 'test/bar.js'];
-		t.deepEqual(actual, expected);
-	} finally {
-		process.chdir(previous);
-	}
-});
-
 test('mergeWithPkgConf: use child if closest', t => {
 	const cwd = path.resolve('fixtures', 'nested', 'child');
 	const result = manager.mergeWithPkgConf({cwd});
